backend/requirements.txt
fastapi[standard]
pydantic_settings
requests
python-dotenv
sqlmodel
openai
google-api-python-client